feat(devices): generate devices keys on the device
Modem firmware 1.3 introduced to ability to generate device
keys on the device, so that the private key never leaves
the device.

This is much more secure and therefore this is now the
recommended way to create certificates.

BREAKING CHANGE: this changes the way credentials for real devices
are generated and support for flashing device private keys from
JSON files has been removed.

Closes #131